backend/requirements.txt
aiohttp==3.9.3
aiohttp-cors==0.7.0
groq==0.4.2
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0